fix: widen ref_entropy_timeseries dtype to uint16
At n >= 256 with picture='purification', the per-layer observable
k = S(R) can exceed 255 and overflowed the uint8 dtype used for the
record_time_series buffer. Widen both the in-memory array and the
on-disk S_of_t dataset to uint16, and update the loader to read as
uint16 regardless of the stored dtype (h5py upcasts automatically).

Fixes OverflowError when running the Fig 7 protocol at n=256 or
n=512 with the full purification picture.

+++ sparsegf2/analysis/single_ref.py
@@ -191,7 +191,10 @@ def load_timeseries(run_dir: Union[str, Path]) -> Dict[Tuple[int, float], CellTi
         except (IndexError, ValueError):
             continue
         with h5py.File(ts_path, "r") as f:
-            S_of_t = np.asarray(f["S_of_t"][:], dtype=np.uint8)
+            # Accept either uint8 (older single_ref data) or uint16
+            # (purification with n > 255). Normalize to uint16 for
+            # downstream analysis.
+            S_of_t = np.asarray(f["S_of_t"][:], dtype=np.uint16)
             t_axis = np.asarray(f["t_axis"][:], dtype=np.int32)
             seeds = np.asarray(f["sample_seed"][:], dtype=np.int64)
         out[(n, p)] = CellTimeseries(
